hrms/urls.py
"""
URL configuration for hrms project.

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/4.2/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.contrib import admin
from django.urls import path, include
from django.urls import re_path as url
from django.conf import settings
from django.views.static import serve
from rest_framework.authtoken.views import obtain_auth_token
from rest_framework_simplejwt import views as jwt_views
from rest_framework import routers
from rest_framework_swagger.views import get_swagger_view

schema_view = get_swagger_view(title='PBZ HRMS REST API')

router = routers.DefaultRouter()

urlpatterns = [
    path('admin/', admin.site.urls),
    path('dictionary/api/', include('dictionary.urls')),
    path('controller/api/', include('controller.urls')),
    path('leave/api/', include('leave.urls')),
    path('payroll/api/', include('payroll.urls')),
    path('training/api/', include('training.urls')),
    path('gateway/api/', include('gateway.urls')),
    path('report/api/', include('report.urls')),
    url(r'^media/(?P<path>.*)$', serve,
        {'document_root': settings.MEDIA_ROOT}),
    url(r'^static/(?P<path>.*)$', serve,
        {'document_root': settings.STATIC_ROOT}),
]

if settings.DEBUG:
    urlpatterns += [
        path('', schema_view),
    ]
