requirements.txt
APScheduler==3.10.4
asgiref==3.5.2
backports.zoneinfo;python_version<"3.9"
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
defusedxml==0.7.1
Django==4.0.6
django-apscheduler==0.6.2
django-cors-headers==3.13.0
django-crontab==0.7.1
django-filter==22.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djangorestframework-xml==2.0.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.1
numpy==1.23.1
openapi-codec==1.3.2
pandas==1.4.3
Pillow==9.2.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.10